requirements.txt
fastapi==0.79.0
pydantic==1.9.2
python-dotenv==0.20.0
SQLAlchemy==1.4.40
